update12262220

+++ src/main/java/com/zftx/mcdaily/bean/TeaChoose.java
@@ -3,6 +3,8 @@ package com.zftx.mcdaily.bean;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 @Data
 @Accessors(chain = true)
 public class TeaChoose {
@@ -10,6 +12,6 @@ public class TeaChoose {
     private Integer userId;
     private Integer teaId;
     private Integer number;
-    private Float money;
+    private Double money;
     private String date;
 }
