+++ src/main/java/com/zftx/mcdaily/controller/TeaRepositoryController.java
@@ -73,7 +73,7 @@ public class TeaRepositoryController {
      */
     @RequestMapping(value = "/addTeaRepository", method = RequestMethod.POST, produces = "application/json;charset=UTF-8")
     @ResponseBody
-    public R addTeaRepository(@RequestParam("catName1")String catName1,@RequestParam("catName2")String catName2,@RequestParam("tName")String tName, MultipartRequest mrequest,@RequestParam("standard") String standard,@RequestParam("price")float price,@RequestParam("note") String note)throws IOException {
+    public R addTeaRepository(@RequestParam("catName1")String catName1,@RequestParam("catName2")String catName2,@RequestParam("tName")String tName, MultipartRequest mrequest,@RequestParam("standard") String standard,@RequestParam("price")Double price,@RequestParam("note") String note)throws IOException {
         TeaRepository teaRepository = new TeaRepository().setTName(tName.trim());
         //茶点名下拉文本切换
         String catName="";
@@ -145,7 +145,7 @@ public class TeaRepositoryController {
      */
     @RequestMapping(value = "/updateTeaRepository", method = RequestMethod.POST, produces = "application/json;charset=UTF-8")
     @ResponseBody
-    public R updateTeaRepository(@RequestParam("id")Integer id,@RequestParam("catName")String catName,@RequestParam("tName")String tName, MultipartRequest mrequest,@RequestParam("standard") String standard,@RequestParam("price")float price,@RequestParam("note") String note)throws IOException{
+    public R updateTeaRepository(@RequestParam("id")Integer id,@RequestParam("catName")String catName,@RequestParam("tName")String tName, MultipartRequest mrequest,@RequestParam("standard") String standard,@RequestParam("price")Double price,@RequestParam("note") String note)throws IOException{
         /*//茶点不能重复
         TeaRepository teaRepository = new TeaRepository().setTName(tName.trim());
         String flag="isNoLike";
